Crate kudu

Structs

AlterTableBuilder

Client
A Kudu database client.

ClientConfig
Client configuration options.

Column
Column instances hold metadata information about columns in a Kudu table.

FlushStats
Carries information about the batches and row operations in a flush.

HashPartitionSchema

Master
Master metadata.

MasterError

MasterId

Partition

PartitionSchema

RangePartitionSchema

Replica
Tablet replica belonging to a tablet server.

ReplicaId

Row

RpcError

Schema

SchemaBuilder

Status

Table

TableBuilder

TableId

Tablet

TabletId

TabletServer
Tablet server metadata.

TabletServerError

TabletServerId

Writer

WriterConfig

Enums

CompressionType

DataType

EncodingType

Error

Event

EventSet
The events that the writer should return in the event channel. Each event set builds on the previous.

MasterErrorCode

OperationType

RaftRole
The role of a Kudu master or tserver in a consensus group.

RangePartitionBound

RpcErrorCode

StatusCode

TabletServerErrorCode

Traits

Value
Marker trait for types which can be stored in a Kudu column.

Type Definitions

Result