Enum kudu::Error

pub enum Error {
    InvalidArgument(String),
    Rpc(RpcError),
    Master(MasterError),
    TabletServer(TabletServerError),
    Io(Error),
    Serialization(String),
    VersionMismatch(String),
    Backoff,
    TimedOut,
    Cancelled,
    ConnectionError,
    NegotiationError(&'static str),
    NoRangePartition,
}

Variants
InvalidArgument(String)

The operation failed because of an invalid argument.

Rpc(RpcError)

A Kudu RPC error.

Master(MasterError)

A Kudu Master error.

TabletServer(TabletServerError)

A Kudu tablet server error.

Io(Error)

An I/O error.

Serialization(String)

An error serializing, deserializing, encoding, or decoding data.

VersionMismatch(String)

An error due to a version mismatch between the client and server.

Backoff

The send queue is full.

TimedOut

The operation timed out. Includes zero or more errors which resulted in retries.

Cancelled

The operation was cancelled.

ConnectionError

The connection encountered an error or hangup.

NegotiationError(&'static str)NoRangePartition

An operation failed because the range partition did not exist.

Methods

impl Error

fn is_network_error(&self) -> bool