Enum kudu::MasterErrorCode

pub enum MasterErrorCode {
    UnknownError,
    InvalidSchema,
    TableNotFound,
    TableAlreadyPresent,
    TooManyTablets,
    CatalogManagerNotInitialized,
    NotTheLeader,
    InvalidReplicationFactor,
    TabletNotRunning,
}

Variants
UnknownError

An error which has no more specific error code. The Status code and message may reveal more details.

InvalidSchema

The schema provided for a request was not well-formed.

TableNotFound

The requested table does not exist

TableAlreadyPresent

The name requested for the table is already in use

TooManyTablets

The number of tablets requested for a new table is over the per TS limit.

CatalogManagerNotInitialized

Catalog manager is not yet initialized.

NotTheLeader

The operation attempted can only be invoked against either the leader or a single non-distributed master, which this node isn't.

InvalidReplicationFactor

The replication factor is invalid, either because it is too low, too high, or an even number.

TabletNotRunning

A tablet involved in the operation is not running.