Enum kudu::TabletServerErrorCode

pub enum TabletServerErrorCode {
    UnknownError,
    InvalidSchema,
    InvalidRowBlock,
    InvalidMutation,
    MismatchedSchema,
    TabletNotFound,
    ScannerExpired,
    InvalidScanSpec,
    InvalidConfig,
    TabletAlreadyExists,
    TabletHasANewerSchema,
    TabletNotRunning,
    InvalidSnapshot,
    InvalidScanCallSeqId,
    NotTheLeader,
    WrongServerUuid,
    CasFailed,
    AlreadyInProgress,
    Throttled,
}

Variants
UnknownError

An error which has no more specific error code. The code and message in 'status' may reveal more details.

InvalidSchema

The schema provided for a request was not well-formed.

InvalidRowBlock

The row data provided for a request was not well-formed.

InvalidMutation

The mutations or mutation keys provided for a request were not well formed.

MismatchedSchema

The schema provided for a request didn't match the actual schema of the tablet.

TabletNotFound

The requested tablet ID is not currently hosted on the server.

ScannerExpired

A request was made against a scanner ID that was either never created or has expired.

InvalidScanSpec

An invalid scan was specified -- e.g the values passed for predicates were incorrect sizes.

InvalidConfig

The provided configuration was not well-formed and/or had a sequence number that was below the current config.

TabletAlreadyExists

On a create tablet request, signals that the tablet already exists.

TabletHasANewerSchema

If the tablet has a newer schema than the requested one the "alter" request will be rejected with this error.

TabletNotRunning

The tablet is hosted on this server, but not in RUNNING state.

InvalidSnapshot

Client requested a snapshot read but the snapshot was invalid.

InvalidScanCallSeqId

An invalid scan call sequence ID was specified.

NotTheLeader

This tserver is not the leader of the consensus configuration.

WrongServerUuid

The destination UUID in the request does not match this server.

CasFailed

The compare-and-swap specified by an atomic RPC operation failed.

AlreadyInProgress

The requested operation is already inprogress, e.g. RemoteBootstrap.

Throttled

The request is throttled.