Struct kudu::Column

pub struct Column { /* fields omitted */ }

Column instances hold metadata information about columns in a Kudu table.

Column also serves as a builder object for specifying new columns during create and alter table operations.

Methods

impl Column

fn name(&self) -> &str

fn data_type(&self) -> DataType

fn is_nullable(&self) -> bool

fn encoding(&self) -> EncodingType

fn compression(&self) -> CompressionType

fn block_size(&self) -> Option<u32>

fn builder<S>(name: S, data_type: DataType) -> Column where S: Into<String>

Returns a new column builder.

fn set_nullable(self) -> Column

fn set_nullable_by_ref(&mut self) -> &mut Column

fn set_not_null(self) -> Column

fn set_not_null_by_ref(&mut self) -> &mut Column

fn set_encoding(self, encoding: EncodingType) -> Column

fn set_encoding_by_ref(&mut self, encoding: EncodingType) -> &mut Column

fn set_compression(self, compression: CompressionType) -> Column

fn set_compression_by_ref(&mut self,
compression: CompressionType)
-> &mut Column

fn set_block_size(self, block_size: u32) -> Column

fn set_block_size_by_ref(&mut self, block_size: u32) -> &mut Column