Struct kudu::TableBuilder

pub struct TableBuilder { /* fields omitted */ }

Methods

impl TableBuilder

fn new<S>(name: S, schema: Schema) -> TableBuilder where S: Into<String>

Creates a new table builder with the provided table name and schema.

fn schema(&self) -> &Schema

fn add_hash_partitions<S>(&mut self,
columns: Vec<S>,
num_partitions: u32)
-> &mut TableBuilder where S: Into<String>

Hash partitions the table by the specfied columns.

fn add_hash_partition_with_seed<S>(&mut self,
columns: Vec<S>,
num_partitions: u32,
seed: u32)
-> &mut TableBuilder where S: Into<String>

fn set_range_partition_columns<S>(&mut self,
columns: Vec<S>)
-> &mut TableBuilder where S: Into<String>

Range partitions the table by the specified columns.

Range partitioned tables must have at least one partition added with TableBuilder::add_range_partition.

fn add_range_partition(&mut self,
lower_bound: RangePartitionBound,
upper_bound: RangePartitionBound)
-> &mut TableBuilder

Adds a range partition to the table with the specified bounds.

fn add_range_partition_split(&mut self, split: Row) -> &mut TableBuilder

fn set_num_replicas(&mut self, num_replicas: u32)